Updated from agora_moodle2 repository (20221006)

+++ html/mod/journal/view.php
@@ -22,18 +22,17 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-require_once("../../config.php");
-require_once("lib.php");
-
+require_once('../../config.php');
+require_once('lib.php');
 
 $id = required_param('id', PARAM_INT);    // Course Module ID.
 
 if (! $cm = get_coursemodule_from_id('journal', $id)) {
-    throw new \moodle_exception(get_string("Course Module ID was incorrect"));
+    throw new \moodle_exception(get_string('incorrectcmid', 'journal'));
 }
 
-if (! $course = $DB->get_record("course", array('id' => $cm->course))) {
-    throw new \moodle_exception(get_string("Course is misconfigured"));
+if (! $course = $DB->get_record('course', array('id' => $cm->course))) {
+    throw new \moodle_exception(get_string('incorrectcourseid', 'journal'));
 }
 
 $context = context_module::instance($cm->id);
@@ -50,12 +49,12 @@ if (!$entriesmanager && !$canadd) {
     throw new \moodle_exception(get_string('accessdenied', 'journal'));
 }
 
-if (! $journal = $DB->get_record("journal", array("id" => $cm->instance))) {
-    throw new \moodle_exception(get_string("Course module is incorrect"));
+if (! $journal = $DB->get_record('journal', array('id' => $cm->instance))) {
+    throw new \moodle_exception(get_string('incorrectjournalid', 'journal'));
 }
 
-if (! $cw = $DB->get_record("course_sections", array("id" => $cm->section))) {
-    throw new \moodle_exception(get_string("Course module is incorrect"));
+if (! $cw = $DB->get_record('course_sections', array('id' => $cm->section))) {
+    throw new \moodle_exception(get_string('incorrectcoursesectionid', 'journal'));
 }
 
 $journalname = format_string($journal->name, true, array('context' => $context));
@@ -72,12 +71,18 @@ echo $OUTPUT->heading($journalname);
 // Check to see if groups are being used here.
 $groupmode = groups_get_activity_groupmode($cm);
 $currentgroup = groups_get_activity_group($cm, true);
+$allowedgroups = groups_get_activity_allowed_groups($cm);
 groups_print_activity_menu($cm, $CFG->wwwroot . "/mod/journal/view.php?id=$cm->id");
 
 if ($entriesmanager) {
-    $entrycount = journal_count_entries($journal, $currentgroup);
-    echo '<div class="reportlink"><a href="report.php?id='.$cm->id.'">'.
-          get_string('viewallentries', 'journal', $entrycount).'</a></div>';
+    if ($currentgroup === 0 && $groupmode == SEPARATEGROUPS && !has_capability('moodle/site:accessallgroups', $context)) {
+        $currentgroup = null;
+    }
+    if (!$currentgroup || array_key_exists($currentgroup, $allowedgroups)) {
+        $entrycount = journal_count_entries($journal, $currentgroup);
+        echo '<div class="reportlink"><a href="report.php?id='.$cm->id.'">'.
+            get_string('viewallentries', 'journal', $entrycount).'</a></div>';
+    }
 }
 
 $journal->intro = trim($journal->intro);
@@ -89,7 +94,7 @@ if (!empty($journal->intro)) {
 echo '<br />';
 
 $timenow = time();
-if ($course->format == 'weeks' and $journal->days) {
+if ($course->format == 'weeks' && $journal->days) {
     $timestart = $course->startdate + (($cw->section - 1) * 604800);
     if ($journal->days) {
         $timefinish = $timestart + (3600 * 24 * $journal->days);
@@ -111,19 +116,19 @@ if ($timenow > $timestart) {
 
         if ($canadd) {
             echo $OUTPUT->single_button('edit.php?id='.$cm->id, get_string('startoredit', 'journal'), 'get',
-                array("class" => "singlebutton journalstart mb-3", "primary" => 1));
+                array('class' => 'singlebutton journalstart mb-3', 'primary' => 1));
         }
     }
 
     // Display entry.
     if ($entry = $DB->get_record('journal_entries', array('userid' => $USER->id, 'journal' => $journal->id))) {
-        echo "<div>";
+        echo '<div>';
         if (empty($entry->text)) {
             echo '<p align="center"><b>'.get_string('blankentry', 'journal').'</b></p>';
         } else {
             echo journal_format_entry_text($entry, $course, $cm);
         }
-        echo "</div>";
+        echo '</div>';
     } else {
         echo '<div><span class="warning">'.get_string('notstarted', 'journal').'</span></div>';
     }
@@ -136,11 +141,11 @@ if ($timenow > $timestart) {
             echo '<div class="lastedit"><strong>'.get_string('lastedited').': </strong> ';
             echo userdate($entry->modified);
             echo ' ('.get_string('numwords', '', count_words($entry->text)).')';
-            echo "</div>";
+            echo '</div>';
         }
         // Added three lines to mark entry as being dirty and needing regrade.
-        if (!empty($entry->modified) AND !empty($entry->timemarked) AND $entry->modified > $entry->timemarked) {
-            echo "<div class=\"lastedit\">".get_string("needsregrade", "journal"). "</div>";
+        if (!empty($entry->modified) && !empty($entry->timemarked) && $entry->modified > $entry->timemarked) {
+            echo '<div class="lastedit">'.get_string('needsregrade', 'journal'). '</div>';
         }
 
         if (!empty($journal->days)) {
@@ -154,7 +159,7 @@ if ($timenow > $timestart) {
     }
 
     // Feedback.
-    if (!empty($entry->entrycomment) or (!empty($entry->rating) and !$entry->rating)) {
+    if (!(empty($entry->entrycomment) || (!empty($entry->rating)) && !$entry->rating)) {
         $grades = make_grades_menu($journal->grade);
         echo $OUTPUT->heading(get_string('feedback'));
         journal_print_feedback($course, $entry, $grades);
